requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.2
image==1.5.33
numpy==1.22.2
opencv-python==4.5.5.62
Pillow==9.0.1
pytesseract==0.3.8
six==1.16.0
sqlparse==0.4.2
